refactored confidential utils
