+++ src/main/kotlin/it/pagopa/ecommerce/helpdesk/utils/v2/SearchParamDecoderV2.kt
@@ -1,14 +1,15 @@
 package it.pagopa.ecommerce.helpdesk.utils.v2
 
-import it.pagopa.ecommerce.helpdesk.utils.ConfidentialMailUtils
+import it.pagopa.ecommerce.helpdesk.utils.confidentialDataUtils
 import it.pagopa.generated.ecommerce.helpdesk.v2.model.HelpDeskSearchTransactionRequestDto
 import it.pagopa.generated.ecommerce.helpdesk.v2.model.SearchTransactionRequestEmailDto
+import it.pagopa.generated.ecommerce.helpdesk.v2.model.SearchTransactionRequestFiscalCodeDto
 import reactor.core.publisher.Mono
 
-/** Data class that wraps search parameter with optional ConfidentialMailUtils class */
+/** Data class that wraps search parameter with optional confidentialDataUtils class */
 class SearchParamDecoderV2<out T>(
     val searchParameter: T,
-    val confidentialMailUtils: ConfidentialMailUtils?
+    val confidentialDataUtils: ConfidentialDataUtils?
 ) where T : HelpDeskSearchTransactionRequestDto {
 
     /**
@@ -20,13 +21,21 @@ class SearchParamDecoderV2<out T>(
     fun decode(): Mono<out HelpDeskSearchTransactionRequestDto> =
         when (searchParameter) {
             /*
-             * Search parameter decoding is performed iff input search parameter is an email and the searching is performed for eCommerce transactions.
-             * In fact confidentialMailUtils parameter is optional and valued only when searching for eCommerce transactions
+             * Search parameter decoding is performed iff input search parameter is an email or fiscal code and the searching is performed for eCommerce transactions.
+             * In fact confidentialDataUtils parameter is optional and valued only when searching for eCommerce transactions
              */
             is SearchTransactionRequestEmailDto ->
-                if (confidentialMailUtils != null) {
-                    confidentialMailUtils.toConfidential(searchParameter.userEmail).map {
+                if (confidentialDataUtils != null) {
+                    confidentialDataUtils.toConfidential(searchParameter.userEmail).map {
                         SearchTransactionRequestEmailDto().userEmail(it.opaqueData).type("EMAIL")
+                    }
+                } else {
+                    Mono.just(searchParameter)
+                }
+            is SearchTransactionRequestFiscalCodeDto ->
+                if (confidentialDataUtils != null) {
+                    confidentialDataUtils.toConfidential(searchParameter.userFiscalCode).map {
+                        SearchTransactionRequestFiscalCodeDto().userFiscalCode(it.opaqueData).type("USER_FISCAL_CODE")
                     }
                 } else {
                     Mono.just(searchParameter)
